Created a function to update the available moves.
Simplified identification of which player's turn it currently is.

Added check to make_move to skip computation if in setup (which should save some time over the life of each game, as there are at most 64 turns)

Next steps: allow the game to be finished and compare this method against a different method of running the game.